MRI/appDatas/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
sqlalchemy==2.0.9
pymysql==1.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pillow==9.5.0
numpy==1.24.3
python-dotenv==1.0.0
aiofiles==23.1.0
websockets==11.0.1
cryptography==40.0.1
bcrypt==4.0.1
configparser==5.3.0
email-validator==2.0.0 